Deprecate `config.Config` and `config.Service`, use `service.Config*`
Fixes https://github.com/open-telemetry/opentelemetry-collector/issues/4605

Followup PRs:
* Move service/internal/configunmarshaler as private struct into service/ to avoid circular dependency (no public API change)
* Change service/internal/builders to private structs in service/ or don't pass config, instead only pass the components and pipeline from the config.

While doing this, I realize that the "service" section in our config is a bit useless, we can think of promoting everything under "service::" one level up.
