+++ config/common.go
@@ -14,6 +14,39 @@
 
 package config // import "go.opentelemetry.io/collector/config"
 
+// Type is the component type as it is used in the config.
+type Type string
+
+// validatable defines the interface for the configuration validation.
+type validatable interface {
+	// Validate validates the configuration and returns an error if invalid.
+	Validate() error
+}
+
+// Unmarshallable defines an optional interface for custom configuration unmarshaling.
+// A configuration struct can implement this interface to override the default unmarshaling.
+type Unmarshallable interface {
+	// Unmarshal is a function that unmarshals a config.Map into the unmarshable struct in a custom way.
+	// The config.Map for this specific component may be nil or empty if no config available.
+	Unmarshal(component *Map) error
+}
+
+// DataType is a special Type that represents the data types supported by the collector. We currently support
+// collecting metrics, traces and logs, this can expand in the future.
+type DataType = Type
+
+// Currently supported data types. Add new data types here when new types are supported in the future.
+const (
+	// TracesDataType is the data type tag for traces.
+	TracesDataType DataType = "traces"
+
+	// MetricsDataType is the data type tag for metrics.
+	MetricsDataType DataType = "metrics"
+
+	// LogsDataType is the data type tag for logs.
+	LogsDataType DataType = "logs"
+)
+
 func unmarshal(componentSection *Map, intoCfg interface{}) error {
 	if cu, ok := intoCfg.(Unmarshallable); ok {
 		return cu.Unmarshal(componentSection)
