+++ config/moved_config_test.go
